Reads an H2O model from file and inserts it back into a previously fitted parsnip model specification or workflow object

H2O models cannot be saved using the typical R approaches such as saveRDS because the actual H2O model is contained within a Java virtual machine. H2O models need to be saved and restored using the `h2o::h2o.saveModel` and `h2o::h2o.loadModel` functions. This is inconvenient for using H2O models contained within parsnip model specifications or workflow objects.

read_h2o(object, filename, ...)

Arguments

object
Either a `workflows::workflow()` object contained a fitted model when using the workflows package, or a `model_spec` object from a fitted model when using the parsnip package directly.

filename
A `character` specifying the file path used to read the model.

Value

The parsnip `model_spec` object or workflows `workflow` object with the H2O model inserted back into the model object.

Details

The `write_h2o` function extracts the H2O model from within a parsnip or workflow fitted model and saves it to file using the `h2o::h2o.saveModel` function. To restore a model and insert it back into a previously fitted model use the `read_h2o` function.

Examples

library(parsnip)
library(h2o)

#> 
#> ----------------------------------------------------------------------
#> 
#> Your next step is to start H2O:
#>     > h2o.init()
#> 
#> For H2O package documentation, ask for help:
#>     > ??h2o
#> 
#> After starting H2O, you can use the Web UI at http://localhost:54321
#> For more information visit http://docs.h2o.ai
#> 
#> ----------------------------------------------------------------------

#> 
#> Attaching package: ‘h2o’

#> The following objects are masked from ‘package:stats’:
#> 
#>     cor, sd, var

#> The following objects are masked from ‘package:base’:
#> 
#>     &&, %*%, %in%, ||, apply, as.factor, as.numeric, colnames,
#>     colnames<-, ifelse, is.character, is.factor, is.numeric, log,
#>     log10, log1p, log2, round, signif, trunc

# start a h2o session
h2o.init()

#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 minutes 9 seconds 
#>     H2O cluster timezone:       America/Edmonton 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.30.0.1 
#>     H2O cluster version age:    2 months and 18 days  
#>     H2O cluster name:           H2O_started_from_R_steven_fie471 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   8.00 GB 
#>     H2O cluster total cores:    16 
#>     H2O cluster allowed cores:  16 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, XGBoost, Algos, AutoML, Core V3, TargetEncoder, Core V4 
#>     R Version:                  R version 4.0.1 (2020-06-06) 
#> 

# fit a parsnip model using the h2o engine
clf <- mlp(mode = "classification") %>%
    set_engine("h2o")

model_fit <- clf %>% fit(Species ~., iris)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

# save the parsnip model
saveRDS(model_fit, file.path(tempdir(), "my_model.rds"))

# save the h2o component of the model
write_h2o(object = model_fit, filename = file.path(tempdir(), "my_h2o_model.mod"))

#> [1] "/var/folders/lc/5k7m4y5j48xc53gx61rcr2p80000gn/T//RtmpnVp1B2/my_h2o_model.mod"

h2o.shutdown(prompt = FALSE)

# restore a model
h2o.init()

#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         3 minutes 11 seconds 
#>     H2O cluster timezone:       America/Edmonton 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.30.0.1 
#>     H2O cluster version age:    2 months and 18 days  
#>     H2O cluster name:           H2O_started_from_R_steven_fie471 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   8.00 GB 
#>     H2O cluster total cores:    16 
#>     H2O cluster allowed cores:  16 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, XGBoost, Algos, AutoML, Core V3, TargetEncoder, Core V4 
#>     R Version:                  R version 4.0.1 (2020-06-06) 
#> 

model_fit <- readRDS(file.path(tempdir(), "my_model.rds"))

# read and insert the H2O model back into the parsnip object
model_fit <- read_h2o(model_fit, file.path(tempdir(), "my_h2o_model.mod"))