Tune h2o models

This is a prototype of a version of tune_grid that uses h2o.grid to perform hyperparameter tuning.

tune_grid_h2o(
  object,
  preprocessor = NULL,
  resamples,
  param_info = NULL,
  grid = 10,
  metric = NULL,
  verbose = FALSE
)

Arguments

object
A parsnip `model_spec` object.

preprocessor
A `recipe` object.

resamples
An `rset` object.

param_info
A `dials::parameters()` object or NULL. If none is given, a parameters set is derived from other arguments. Passing this argument can be useful when parameter ranges need to be customized.

grid
A `data.frame` of tuning combinations or a positive integer. The data frame should have columns for each parameter being tuned and rows for tuning parameter candidates. An integer denotes the number of candidate parameter sets to be created automatically. If a positive integer is used or no tuning grid is supplied, then a semi-random grid via `dials::grid_latin_hypercube` is created based on the specified number of tuning iterations (default size = 10).

metric
A `character` vector specifying the h2o metric to use as the scoring criterion for the hyperparameter tuning. Must be one of c("rsq", "sensitivity", "rmse", "accuracy", "mn_log_loss", "mse"). The default is "rsq" for regression models and "mn_log_loss" for classification models.

verbose
Whether to provide progress messages.

Value

Limitations

- Only model arguments can be tuned, not arguments in the preprocessing recipes.

- Parsnip only allows `data.frame` and `tbl_spark` objects to be passed to the `fit` method, not `H2OFrame` objects.